+++ examples/usual/dynamic_options/format/duration/message/lambda/example2.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+module Usual
+  module DynamicOptions
+    module Format
+      module Duration
+        module Message
+          module Lambda
+            class Example2 < ApplicationService::Base
+              input :song_duration, type: String
+
+              internal :song_duration,
+                       type: String,
+                       check_format: {
+                         is: :duration,
+                         message: lambda do |internal:, value:, option_value:, **|
+                           "Value `#{value}` does not match the format of `#{option_value}` in `#{internal.name}`"
+                         end
+                       }
+
+              output :song_duration, type: ActiveSupport::Duration
+
+              make :assign_internal
+
+              make :assign_output
+
+              private
+
+              def assign_internal
+                internals.song_duration = inputs.song_duration
+              end
+
+              def assign_output
+                outputs.song_duration = ActiveSupport::Duration.parse(internals.song_duration)
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end