+++ lib/generators/servactory/templates/services/application_service/base.rb
@@ -53,6 +53,8 @@ module ApplicationService
       # action_shortcuts %i[assign build create save]
 
       # action_aliases %i[do_it!]
+
+      # predicate_methods_enabled false
     end
   end
 end