+++ lib/servactory/context/workspace/internals.rb
@@ -53,15 +53,15 @@ module Servactory
           @context.send(:servactory_service_store).assign_internal(internal.name, value)
         end
 
-        def getter_with(name:, &block) # rubocop:disable Lint/UnusedMethodArgument
-          internal_name = name.to_s.chomp("?").to_sym
+        def getter_with(name:, &block) # rubocop:disable Metrics/AbcSize, Lint/UnusedMethodArgument
+          internal_name = @context.class.config.predicate_methods_enabled? ? name.to_s.chomp("?").to_sym : name
           internal = @collection_of_internals.find_by(name: internal_name)
 
           return yield if internal.nil?
 
           internal_value = @context.send(:servactory_service_store).fetch_internal(internal.name)
 
-          if name.to_s.end_with?("?")
+          if name.to_s.end_with?("?") && @context.class.config.predicate_methods_enabled?
             Servactory::Utils.query_attribute(internal_value)
           else
             internal_value