+++ lib/servactory/test_kit/rspec/matchers/have_service_input_matcher.rb
@@ -1,5 +1,10 @@
 # frozen_string_literal: true
 
+require_relative "have_service_input_matchers/default_matcher"
+require_relative "have_service_input_matchers/optional_matcher"
+require_relative "have_service_input_matchers/required_matcher"
+require_relative "have_service_input_matchers/valid_with_matcher"
+
 module Servactory
   module TestKit
     module Rspec