+++ lib/servactory/tool_kit/dynamic_options/format.rb
@@ -3,15 +3,31 @@
 module Servactory
   module ToolKit
     module DynamicOptions
-      class Format < Must
+      class Format < Must # rubocop:disable Metrics/ClassLength
         DEFAULT_FORMATS = {
-          boolean: {
-            pattern: /^(true|false|0|1)$/i,
-            validator: ->(value:) { %w[true 1].include?(value&.downcase) }
+          uuid: {
+            pattern: /^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$/,
+            validator: ->(value:) { value.present? }
           },
           email: {
             pattern: URI::MailTo::EMAIL_REGEXP,
             validator: ->(value:) { value.present? }
+          },
+          password: {
+            # NOTE: Pattern 4 » https://dev.to/rasaf_ibrahim/write-regex-password-validation-like-a-pro-5175
+            #       Password must contain one digit from 1 to 9, one lowercase letter, one
+            #       uppercase letter, and one underscore, and it must be 8-16 characters long.
+            #       Usage of any other special character and usage of space is optional.
+            pattern: /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z]).{8,16}$/,
+            validator: ->(value:) { value.present? }
+          },
+          duration: {
+            pattern: nil,
+            validator: lambda do |value:|
+              ActiveSupport::Duration.parse(value) and return true
+            rescue ActiveSupport::Duration::ISO8601Parser::ParsingError
+              false
+            end
           },
           date: {
             pattern: nil,
@@ -29,14 +45,6 @@ module Servactory
               false
             end
           },
-          password: {
-            # NOTE: Pattern 4 » https://dev.to/rasaf_ibrahim/write-regex-password-validation-like-a-pro-5175
-            #       Password must contain one digit from 1 to 9, one lowercase letter, one
-            #       uppercase letter, and one underscore, and it must be 8-16 characters long.
-            #       Usage of any other special character and usage of space is optional.
-            pattern: /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z]).{8,16}$/,
-            validator: ->(value:) { value.present? }
-          },
           time: {
             pattern: nil,
             validator: lambda do |value:|
@@ -44,6 +52,10 @@ module Servactory
             rescue ArgumentError
               false
             end
+          },
+          boolean: {
+            pattern: /^(true|false|0|1)$/i,
+            validator: ->(value:) { %w[true 1].include?(value&.downcase) }
           }
         }.freeze
         private_constant :DEFAULT_FORMATS