+++ spec/examples/wrong/prepare/example2_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Wrong::Prepare::Example3, type: :service do
+RSpec.describe Wrong::Prepare::Example2, type: :service do
   describe ".call!" do
     subject(:perform) { described_class.call!(**attributes) }